requirements.txt
fastapi<1
uvicorn<1
slack_bolt==1.18.0

aiohttp==3.8.5
brotli==1.1.0
aiodns==3.0.0
pycares==4.3.0

pymongo==4.5.0
dnspython==2.4.2

python-dotenv==1.0.0
langchain==0.0.285
openai==0.28.0
pyyaml==6.0.*